requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
Deprecated==1.3.1
distro==1.9.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
img2pdf==0.6.3
jiter==0.14.0
lxml==6.1.0
numpy==2.4.4
openai==2.33.0
opencv-python==4.13.0.92
packaging==26.2
pdf2image==1.17.0
pikepdf==10.5.1
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
pytesseract==0.3.13
python-dotenv==1.2.2
python-multipart==0.0.27
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
wrapt==2.1.2
